UrlShortener: update controllers, dtos, repositories,services

+++ UrlShortener/src/main/java/de/telran/UrlShortener/dtos/StatisticClickedUrlRequestDto.java
@@ -12,8 +12,8 @@ import java.util.List;
 public class StatisticClickedUrlRequestDto {
     private Long userId;
     // if periodDays = 0 - get overall, else clickedAmount for the period
-    private Integer periodDays;
-    private Integer limitTop;
+    private Long periodDays;
+    private Long limitTop;
     // if descent = true - non-popular first
     private Boolean descent;
